fixed normalisation & use doubles instead of floats

+++ Context.cs
@@ -22,7 +22,7 @@ namespace INFDTA01_1
 		/// <summary>
 		/// Compute the similarity using the interface.
 		/// </summary>
-		public SortedDictionary<int, double> Compute(SortedDictionary<int, float> targetUser, SortedDictionary<int, SortedDictionary<int, float>> userItems)
+		public SortedDictionary<int, double> Compute(SortedDictionary<int, double> targetUser, SortedDictionary<int, SortedDictionary<int, double>> userItems)
 		{
             return this._similarityInterface.Compute(targetUser, userItems);
 		}
