+++ Strategy/SimilarityInterface.cs
@@ -5,6 +5,6 @@ namespace INFDTA01_1.Strategy
 {
 	public interface SimilarityInterface
 	{
-		SortedDictionary<int, double> Compute(SortedDictionary<int, float> targetUser, SortedDictionary<int, SortedDictionary<int, float>> userItems);
+		SortedDictionary<int, double> Compute(SortedDictionary<int, double> targetUser, SortedDictionary<int, SortedDictionary<int, double>> userItems);
 	}
 }
